pyproject.toml
[tool.poetry]
name = "function"
version = "0.1.0"
description = ""
authors = ["AnasHagras"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"
django-model-utils = "^4.2.0"
django-rest-knox = "^4.2.0"
django-filter = "^22.1"
celery = {extras = ["redis"], version = "^5.2.7"}
redis = "^4.3.4"
django-celery-results = "^2.4.0"
django-celery-beat = "^2.4.0"
django-rest-auth = "^0.9.5"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
